+++ test/specs/unit/querySpec.js
@@ -712,7 +712,10 @@ describe("The Query Class", function() {
                 success: true,
                 data: [{id: "a"}],
                 xhr: {
-                    getResponseHeader: function() {return 6;},
+                     getResponseHeader: function(name) {
+                         if (name == 'Layer-Count') return 6;
+                         if (name == 'Layer-Conversation-Is-Syncing') return 'false';
+                     }
                 }
             }, requestUrl);
             expect(query._appendResults).toHaveBeenCalledWith({
@@ -722,7 +725,7 @@ describe("The Query Class", function() {
             });
         });
 
-        it("Should not call _run if reached the end of the server's results", function() {
+        it("Should not schedule call _run if results are up to date", function() {
             spyOn(query, "_run");
             spyOn(query, "_appendResults");
             query.paginationWindow = 100;
@@ -731,10 +734,34 @@ describe("The Query Class", function() {
                 success: true,
                 data: [{id: "a"}],
                 xhr: {
-                    getResponseHeader: function() {return 6;},
+                     getResponseHeader: function(name) {
+                         if (name == 'Layer-Count') return 6;
+                         if (name == 'Layer-Conversation-Is-Syncing') return 'false';
+                     }
                 }
             }, requestUrl);
+            jasmine.clock().tick(10000);
             expect(query._run).not.toHaveBeenCalled();
+            query.data = [];
+        });
+
+        it("Should schedule _run if results are not up to date", function() {
+            spyOn(query, "_run");
+            spyOn(query, "_appendResults");
+            query.paginationWindow = 100;
+            query.data = [message, message, message, message, message, message, message];
+            query._processRunResults({
+                success: true,
+                data: [{id: "a"}],
+                xhr: {
+                     getResponseHeader: function(name) {
+                         if (name == 'Layer-Count') return 6;
+                         if (name == 'Layer-Conversation-Is-Syncing') return 'true';
+                     }
+                }
+            }, requestUrl);
+            jasmine.clock().tick(10000);
+            expect(query._run).toHaveBeenCalled();
             query.data = [];
         });
 
@@ -744,7 +771,12 @@ describe("The Query Class", function() {
             query._processRunResults({
                 success: true,
                 data: [{id: "a"}],
-                xhr: {}
+                xhr: {
+                     getResponseHeader: function(name) {
+                         if (name == 'Layer-Count') return 6;
+                         if (name == 'Layer-Conversation-Is-Syncing') return 'false';
+                     }
+                }
             }, 'joe');
             expect(query._appendResults).not.toHaveBeenCalled();
         });
@@ -755,7 +787,12 @@ describe("The Query Class", function() {
             query._processRunResults({
                 success: true,
                 data: [{id: "a"}],
-                xhr: {}
+                xhr: {
+                     getResponseHeader: function(name) {
+                         if (name == 'Layer-Count') return 6;
+                         if (name == 'Layer-Conversation-Is-Syncing') return 'false';
+                     }
+                }
             }, 'joe');
             expect(query._appendResults).not.toHaveBeenCalled();
 
@@ -770,7 +807,12 @@ describe("The Query Class", function() {
             query._processRunResults({
                 success: true,
                 data: [{id: "a"}],
-                xhr: {}
+                xhr: {
+                     getResponseHeader: function(name) {
+                         if (name == 'Layer-Count') return 6;
+                         if (name == 'Layer-Conversation-Is-Syncing') return 'false';
+                     }
+                }
             }, 'joe');
             expect(query.isFiring).toBe(true);
             expect(query._firingRequest).toEqual('fred');
@@ -781,66 +823,13 @@ describe("The Query Class", function() {
                 success: true,
                 data: [responses.message1],
                 xhr: {
-                    getResponseHeader: function() {return 6;},
+                     getResponseHeader: function(name) {
+                         if (name == 'Layer-Count') return 6;
+                         if (name == 'Layer-Conversation-Is-Syncing') return 'false';
+                     }
                 }
             }, requestUrl);
             expect(query.totalSize).toEqual(6);
-        });
-
-        it("Should retry up to MaxRetryCount=10 times and then trigger a data event if no data", function() {
-            layer.Query.MaxRetryCount = 10;
-            query.data = [];
-            spyOn(query, "_triggerChange");
-            spyOn(query, "_run").and.callFake(function() {
-               query._firingRequest = requestUrl;
-               query._processRunResults({
-                 success: true,
-                    data: [],
-                    xhr: {
-                        getResponseHeader: function() {return 0;},
-                    }
-                }, requestUrl);
-                jasmine.clock().tick(1600);
-            });
-
-            query._processRunResults({
-                success: true,
-                data: [],
-                xhr: {
-                    getResponseHeader: function() {return 0;},
-                }
-            }, requestUrl);
-            jasmine.clock().tick(1600);
-            expect(query._run.calls.count()).toEqual(10);
-            expect(query._triggerChange).toHaveBeenCalled();
-            layer.Query.MaxRetryCount = 0;
-        });
-
-        it("Should retry up to MaxRetryCount=DEFAULT=0 times and then trigger a data event if no data", function() {
-            query.data = [];
-            spyOn(query, "_triggerChange");
-            spyOn(query, "_run").and.callFake(function() {
-               query._firingRequest = requestUrl;
-               query._processRunResults({
-                 success: true,
-                    data: [],
-                    xhr: {
-                        getResponseHeader: function() {return 0;},
-                    }
-                }, requestUrl);
-                jasmine.clock().tick(1600);
-            });
-
-            query._processRunResults({
-                success: true,
-                data: [],
-                xhr: {
-                    getResponseHeader: function() {return 0;},
-                }
-            }, requestUrl);
-            jasmine.clock().tick(1600);
-            expect(query._run.calls.count()).toEqual(0);
-            expect(query._triggerChange).toHaveBeenCalled();
         });
     });
 
@@ -861,7 +850,15 @@ describe("The Query Class", function() {
         it("Should register new results", function() {
             spyOn(client, "_createObject");
             spyOn(client, "_getObject").and.returnValue(conversation);
-            query._appendResults({data: [JSON.parse(JSON.stringify(responses.conversation2))]});
+            query._appendResults({
+                data: [JSON.parse(JSON.stringify(responses.conversation2))],
+                xhr: {
+                     getResponseHeader: function(name) {
+                         if (name == 'Layer-Count') return 6;
+                         if (name == 'Layer-Conversation-Is-Syncing') return 'false';
+                     }
+                }
+            });
             expect(client._createObject).toHaveBeenCalledWith(responses.conversation2);
         });
 
@@ -871,7 +868,15 @@ describe("The Query Class", function() {
             conversation.createdAt.setHours(conversation.createdAt.getHours() + 1);
 
             // Run
-            query._appendResults({data: [JSON.parse(JSON.stringify(responses.conversation2))]});
+            query._appendResults({
+                data: [JSON.parse(JSON.stringify(responses.conversation2))],
+                xhr: {
+                     getResponseHeader: function(name) {
+                         if (name == 'Layer-Count') return 6;
+                         if (name == 'Layer-Conversation-Is-Syncing') return 'false';
+                     }
+                }
+            });
 
             // Posttest
             expect(query.data).not.toBe(oldData);
@@ -890,7 +895,15 @@ describe("The Query Class", function() {
             var oldData = query.data = [conversation];
 
             // Run
-            query._appendResults({data: [JSON.parse(JSON.stringify(responses.conversation2))]});
+            query._appendResults({
+                data: [JSON.parse(JSON.stringify(responses.conversation2))],
+                xhr: {
+                     getResponseHeader: function(name) {
+                         if (name == 'Layer-Count') return 6;
+                         if (name == 'Layer-Conversation-Is-Syncing') return 'false';
+                     }
+                }
+            });
 
             // Posttest
             expect(query.data).toBe(oldData);
@@ -899,7 +912,15 @@ describe("The Query Class", function() {
 
         it("Should put objects rather than instances if dataType is object", function() {
             query.dataType = "object";
-            query._appendResults({data: [JSON.parse(JSON.stringify(responses.conversation2))]});
+            query._appendResults({
+                data: [JSON.parse(JSON.stringify(responses.conversation2))],
+                xhr: {
+                     getResponseHeader: function(name) {
+                         if (name == 'Layer-Count') return 6;
+                         if (name == 'Layer-Conversation-Is-Syncing') return 'false';
+                     }
+                }
+            });
             expect(query.data[0] instanceof layer.Conversation).toBe(false);
         });
 
@@ -919,7 +940,15 @@ describe("The Query Class", function() {
           }).and.returnValue(1);
 
           // Run
-          query._appendResults({data: [c3]});
+          query._appendResults({
+              data: [c3],
+              xhr: {
+                    getResponseHeader: function(name) {
+                        if (name == 'Layer-Count') return 6;
+                        if (name == 'Layer-Conversation-Is-Syncing') return 'false';
+                    }
+              }
+            });
 
           // Posttest
           expect(query._getInsertConversationIndex).toHaveBeenCalled();